refactor/ models to schema and sql to models
