+++ mealie/db/database.py
@@ -1,11 +1,11 @@
 from sqlalchemy.orm.session import Session
 
 from db.db_base import BaseDocument
-from db.sql.meal_models import MealPlanModel
-from db.sql.recipe_models import Category, RecipeModel, Tag
-from db.sql.settings_models import SiteSettingsModel
-from db.sql.theme_models import SiteThemeModel
-from db.sql.users import User
+from db.models.meal_models import MealPlanModel
+from db.models.recipe_models import Category, RecipeModel, Tag
+from db.models.settings_models import SiteSettingsModel
+from db.models.theme_models import SiteThemeModel
+from db.models.users import User
 
 """
 # TODO
@@ -55,10 +55,12 @@ class _Themes(BaseDocument):
         self.primary_key = "name"
         self.sql_model = SiteThemeModel
 
+
 class _Users(BaseDocument):
     def __init__(self) -> None:
         self.primary_key = "id"
         self.sql_model = User
+
 
 class Database:
     def __init__(self) -> None:
