+++ mealie/db/db_base.py
@@ -3,7 +3,7 @@ from typing import List
 from sqlalchemy.orm import load_only
 from sqlalchemy.orm.session import Session
 
-from db.sql.model_base import SqlAlchemyBase
+from db.models.model_base import SqlAlchemyBase
 
 
 class BaseDocument:
