+++ mealie/db/models/theme_models.py
@@ -1,6 +1,6 @@
 import sqlalchemy as sa
 import sqlalchemy.orm as orm
-from db.sql.model_base import BaseMixins, SqlAlchemyBase
+from db.models.model_base import BaseMixins, SqlAlchemyBase
 
 
 class SiteThemeModel(SqlAlchemyBase):
