--- mealie/db/sql/_all_models.py
@@ -1,5 +1,0 @@
-from db.sql.meal_models import *
-from db.sql.recipe_models import *
-from db.sql.settings_models import *
-from db.sql.theme_models import *
-from db.sql.users import *
