+++ mealie/schema/settings_models.py
