+++ mealie/schema/snackbar.py
