+++ mealie/services/backups/exports.py
@@ -9,7 +9,7 @@ from db.db_setup import create_session
 from jinja2 import Template
 from services.meal_services import MealPlan
 from services.recipe_services import Recipe
-from utils.logger import logger
+from fastapi.logger import logger
 
 
 class ExportDatabase:
