+++ mealie/services/backups/imports.py
@@ -6,11 +6,11 @@ from typing import List
 
 from core.config import BACKUP_DIR, IMG_DIR, TEMP_DIR
 from db.database import db
-from models.import_models import RecipeImport, SettingsImport, ThemeImport
-from models.theme_models import SiteTheme
+from schema.import_models import RecipeImport, SettingsImport, ThemeImport
+from schema.theme_models import SiteTheme
 from services.recipe_services import Recipe
 from sqlalchemy.orm.session import Session
-from utils.logger import logger
+from fastapi.logger import logger
 
 
 class ImportDatabase:
