+++ mealie/services/scraper/scraper.py
@@ -8,7 +8,7 @@ from services.image_services import scrape_image
 from services.recipe_services import Recipe
 from services.scraper import open_graph
 from services.scraper.cleaner import Cleaner
-from utils.logger import logger
+from fastapi.logger import logger
 
 LAST_JSON = DEBUG_DIR.joinpath("last_recipe.json")
